backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg[binary]
pydantic-settings
python-dotenv